Feat/be#289: 마이페이지 네비·탈퇴(개인정보)·헤더 탭·요약 제거
Made-with: Cursor
