+++ frontend/src/pages/MypageProfilePage.jsx
@@ -1,35 +1,21 @@
 import { useEffect, useMemo, useState } from 'react'
-import { Link, useNavigate } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 
 import { MypageDevHint } from '../components/MypageDevHint.jsx'
-import { apiRequest } from '../api/client.js'
 import { useAuth } from '../context/useAuth.js'
 import { getGuestDisplayName, loadGuestPrivacyForm } from '../lib/guestMypagePrefs.js'
 import { getEmailFromToken, getRoleFromToken, parseJwtPayload } from '../lib/jwt.js'
 
 import './MypageMemberPages.css'
 
-function parseApiErrorMessage(text) {
-  if (!text) return '요청 실패'
-  try {
-    const j = JSON.parse(text)
-    return (j.message ?? text) || '요청 실패'
-  } catch {
-    return text || '요청 실패'
-  }
-}
-
 export function MypageProfilePage() {
-  const { email, token, logout } = useAuth()
-  const navigate = useNavigate()
+  const { email, token } = useAuth()
 
   const claims = useMemo(() => parseJwtPayload(token ?? ''), [token])
   const jwtEmail = useMemo(() => (token ? getEmailFromToken(token) : null), [token])
   const jwtRole = useMemo(() => (token ? getRoleFromToken(token) : null), [token])
 
   const [nickname, setNickname] = useState('')
-  const [busy, setBusy] = useState(false)
-  const [toast, setToast] = useState('')
 
   useEffect(() => {
     const f = loadGuestPrivacyForm(email)
@@ -37,25 +23,6 @@ export function MypageProfilePage() {
   }, [email])
 
   const displayName = useMemo(() => getGuestDisplayName(email), [email, nickname])
-
-  const onWithdraw = async () => {
-    const ok = window.confirm('정말 탈퇴할까요? 탈퇴 후에는 복구가 어려울 수 있어요.')
-    if (!ok) return
-
-    setBusy(true)
-    setToast('')
-    try {
-      const res = await apiRequest('/members/me?role=GUEST', { method: 'DELETE' })
-      const text = await res.text()
-      if (!res.ok) throw new Error(parseApiErrorMessage(text))
-      await logout()
-      navigate('/', { replace: true, state: { hint: '회원 탈퇴가 완료되었습니다.' } })
-    } catch (e) {
-      setToast(e instanceof Error ? e.message : '탈퇴 처리 실패')
-    } finally {
-      setBusy(false)
-    }
-  }
 
   return (
     <div className="mp-member">
@@ -67,8 +34,6 @@ export function MypageProfilePage() {
       <MypageDevHint>
         회원 API: <code>POST /members/join</code>, <code>DELETE /members/me?role=...</code> · 표시는 JWT + 로컬 prefs
       </MypageDevHint>
-
-      {toast && <p className="err">{toast}</p>}
 
       <div className="mp-scrap-hero" style={{ marginBottom: '1rem' }}>
         <div>
@@ -104,14 +69,6 @@ export function MypageProfilePage() {
             </Link>
           </div>
         </article>
-      </div>
-
-      <h2 style={{ margin: '1.5rem 0 0.75rem', fontSize: '0.95rem', fontWeight: 800 }}>위험 구역</h2>
-      <p className="sub">탈퇴는 되돌리기 어려운 동작입니다. 로컬 개발 계정에서만 테스트해 주세요.</p>
-      <div style={{ display: 'flex', flexWrap: 'wrap', gap: '0.45rem' }}>
-        <button type="button" className="mp-btn mp-btn--danger" disabled={busy} onClick={() => void onWithdraw()}>
-          {busy ? '처리 중…' : '회원 탈퇴(DELETE /members/me?role=GUEST)'}
-        </button>
       </div>
     </div>
   )
